pyproject.toml
[project]
name = "microblog-nostr-bridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "nostr-sdk>=0.44.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
